Fix 'stream' circular dependency...
Import 'pouch-websocket-sync' leads to requiring 'stream'.
In the browser, 'stream' is supplied by 'node-libs-browser'
requiring 'stream-browserify'.

Unfortunately, 'stream-browserify' ultimately requires 'stream' again.
(Presumably this is to inherit from a built-in definition when the
runtime environment provides one.)
When webpack encounters this circle, it eventually returns an empty object,
supplying an incomplete dependency.
This appears to be how CommonJS is supposed to work.

Here the circle is avoided by importing 'stream-browserify' before
'node-libs-browser' closes the circle on 'stream'.

Importing 'stream-browserify' early effectively imports it with a
smaller context where 'stream' is undefined.
The 'readable-stream' code is written to handle that case, knowing that
'stream-browserify' actually is providing that definition in some cases.

This problem seems similar to one reported as [readable-stream issue 237]
(https://github.com/nodejs/readable-stream/issues/237).
That was closed, indicating user error.

[webpack issue 272](https://github.com/webpack/webpack/issues/272)
has info from many others encountering problems with circular dependencies
in webpack.

[webpack issue 631](https://github.com/webpack/webpack/issues/631)
contains a higher level discussion of how the webpack behavior
causes hard to identify runtime failures.
It also indicates that webpack behaves the same as node,
with many folks calling for some sort of warning.
The issue was closed with reference to the [circular-dependency-plugin]
(https://github.com/aackerman/circular-dependency-plugin).

The stack overflow article [circular imports with webpack returning empty object]
(http://stackoverflow.com/questions/30378226/circular-imports-with-webpack-returning-empty-object)
indicates that this behavior is part of CommonJS.

None of the webpack references mention 'stream' as a
source of the circular dependency.
In fact, they imply that this is not an unusual problem.