Rewrite tests using shallow rendering and simplified descriptions...
The important code tests the handlers for logic errors.
Shallow rendering provides access to handlers and 100% coverage.

Testing `props` documents the component API.
Without this, they would have to be mapped out of code coverage.

There is little value in checking that React renders reliably.
No need to test in a DOM.